Clean up code
Note: Unless there's a very specific way that the data are intended to
be processed, it seems to make sense to just load all the data into
memory at once (with a call to .values), otherwise using xarray could
actually end up being slower if the same data segments are reloaded
multiple times. This is the same thing that happens when the data are
loaded with netCDF4.MFDataset (and referenced with [:]).