Add iplot() interactive plot function
Notes / implementation issues:
* Weird memory and plotting issue. If allow selection of field name on
the fly, the the resulting field plot is not guaranteed to plot the
right field... moreover, every time the field changes, a new plot is
leftover in the notebook cell output. This is believed to be a memory
issue because if the field var 'F' to be plotted is set to
self.field[name][itime,:,:] vs self.field[name], the number of residual
plots changes...
* Attempted to use plt.pcolormesh followed by QuadSet.set_array() to
update the field, but unclear how to rescale (renorm) the data (i.e.,
how to set vmin/vmax).